+++ mm2src/coins/nft.rs
@@ -1,19 +1,21 @@
 use mm2_core::mm_ctx::MmArc;
 use mm2_err_handle::prelude::{MmError, MmResult};
+use std::str::FromStr;
 
 pub(crate) mod nft_errors;
 pub(crate) mod nft_structs;
 #[cfg(any(test, target_arch = "wasm32"))] mod nft_tests;
 
 use crate::WithdrawError;
-use nft_errors::GetNftInfoError;
+use nft_errors::{GetInfoFromUriError, GetNftInfoError};
 use nft_structs::{ConvertChain, Nft, NftList, NftListReq, NftMetadataReq, NftTransferHistory,
                   NftTransferHistoryWrapper, NftTransfersReq, NftWrapper, NftsTransferHistoryList,
                   TransactionNftDetails, WithdrawNftReq};
 
 use crate::eth::{get_eth_address, withdraw_erc1155, withdraw_erc721};
-use crate::nft::nft_structs::WithdrawNftType;
+use crate::nft::nft_structs::{TransferStatus, UriMeta, WithdrawNftType};
 use common::APPLICATION_JSON;
+use ethereum_types::Address;
 use http::header::ACCEPT;
 use mm2_err_handle::map_to_mm::MapToMmResult;
 use mm2_number::BigDecimal;
@@ -52,10 +54,11 @@ pub async fn get_nft_list(ctx: MmArc, req: NftListReq) -> MmResult<NftList, GetN
         let mut cursor = String::new();
         loop {
             let uri = format!("{}{}", uri_without_cursor, cursor);
-            let response = send_moralis_request(uri.as_str()).await?;
+            let response = send_request_to_uri(uri.as_str()).await?;
             if let Some(nfts_list) = response["result"].as_array() {
                 for nft_json in nfts_list {
                     let nft_wrapper: NftWrapper = serde_json::from_str(&nft_json.to_string())?;
+                    let uri_meta = try_get_uri_meta(&nft_wrapper.token_uri).await?;
                     let nft = Nft {
                         chain,
                         token_address: nft_wrapper.token_address,
@@ -66,7 +69,7 @@ pub async fn get_nft_list(ctx: MmArc, req: NftListReq) -> MmResult<NftList, GetN
                         block_number_minted: *nft_wrapper.block_number_minted,
                         block_number: *nft_wrapper.block_number,
                         contract_type: nft_wrapper.contract_type.map(|v| v.0),
-                        name: nft_wrapper.name,
+                        collection_name: nft_wrapper.name,
                         symbol: nft_wrapper.symbol,
                         token_uri: nft_wrapper.token_uri,
                         metadata: nft_wrapper.metadata,
@@ -74,6 +77,7 @@ pub async fn get_nft_list(ctx: MmArc, req: NftListReq) -> MmResult<NftList, GetN
                         last_metadata_sync: nft_wrapper.last_metadata_sync,
                         minter_address: nft_wrapper.minter_address,
                         possible_spam: nft_wrapper.possible_spam,
+                        uri_meta,
                     };
                     // collect NFTs from the page
                     res_list.push(nft);
@@ -115,8 +119,9 @@ pub async fn get_nft_metadata(_ctx: MmArc, req: NftMetadataReq) -> MmResult<Nft,
         .append_pair(MORALIS_FORMAT_QUERY_NAME, MORALIS_FORMAT_QUERY_VALUE);
     drop_mutability!(uri);
 
-    let response = send_moralis_request(uri.as_str()).await?;
+    let response = send_request_to_uri(uri.as_str()).await?;
     let nft_wrapper: NftWrapper = serde_json::from_str(&response.to_string())?;
+    let uri_meta = try_get_uri_meta(&nft_wrapper.token_uri).await?;
     let nft_metadata = Nft {
         chain: req.chain,
         token_address: nft_wrapper.token_address,
@@ -127,7 +132,7 @@ pub async fn get_nft_metadata(_ctx: MmArc, req: NftMetadataReq) -> MmResult<Nft,
         block_number_minted: *nft_wrapper.block_number_minted,
         block_number: *nft_wrapper.block_number,
         contract_type: nft_wrapper.contract_type.map(|v| v.0),
-        name: nft_wrapper.name,
+        collection_name: nft_wrapper.name,
         symbol: nft_wrapper.symbol,
         token_uri: nft_wrapper.token_uri,
         metadata: nft_wrapper.metadata,
@@ -135,6 +140,7 @@ pub async fn get_nft_metadata(_ctx: MmArc, req: NftMetadataReq) -> MmResult<Nft,
         last_metadata_sync: nft_wrapper.last_metadata_sync,
         minter_address: nft_wrapper.minter_address,
         possible_spam: nft_wrapper.possible_spam,
+        uri_meta,
     };
     Ok(nft_metadata)
 }
@@ -164,12 +170,22 @@ pub async fn get_nft_transfers(ctx: MmArc, req: NftTransfersReq) -> MmResult<Nft
 
         // The cursor returned in the previous response (used for getting the next page).
         let mut cursor = String::new();
+        let wallet_address = my_address.wallet_address;
         loop {
             let uri = format!("{}{}", uri_without_cursor, cursor);
-            let response = send_moralis_request(uri.as_str()).await?;
+            let response = send_request_to_uri(uri.as_str()).await?;
             if let Some(transfer_list) = response["result"].as_array() {
                 for transfer in transfer_list {
                     let transfer_wrapper: NftTransferHistoryWrapper = serde_json::from_str(&transfer.to_string())?;
+                    let status = get_tx_status(&wallet_address, &transfer_wrapper.to_address);
+                    let req = NftMetadataReq {
+                        token_address: Address::from_str(&transfer_wrapper.token_address)
+                            .map_to_mm(|e| GetNftInfoError::AddressError(e.to_string()))?,
+                        token_id: transfer_wrapper.token_id.clone(),
+                        chain,
+                        url: req.url.clone(),
+                    };
+                    let nft_meta = get_nft_metadata(ctx.clone(), req).await?;
                     let transfer_history = NftTransferHistory {
                         chain,
                         block_number: *transfer_wrapper.block_number,
@@ -183,8 +199,12 @@ pub async fn get_nft_transfers(ctx: MmArc, req: NftTransfersReq) -> MmResult<Nft
                         transaction_type: transfer_wrapper.transaction_type,
                         token_address: transfer_wrapper.token_address,
                         token_id: transfer_wrapper.token_id.0,
+                        collection_name: nft_meta.collection_name,
+                        image: nft_meta.uri_meta.image,
+                        token_name: nft_meta.uri_meta.token_name,
                         from_address: transfer_wrapper.from_address,
                         to_address: transfer_wrapper.to_address,
+                        status,
                         amount: transfer_wrapper.amount.0,
                         verified: transfer_wrapper.verified,
                         operator: transfer_wrapper.operator,
@@ -222,7 +242,7 @@ pub async fn withdraw_nft(ctx: MmArc, req: WithdrawNftReq) -> WithdrawNftResult 
 }
 
 #[cfg(not(target_arch = "wasm32"))]
-async fn send_moralis_request(uri: &str) -> MmResult<Json, GetNftInfoError> {
+async fn send_request_to_uri(uri: &str) -> MmResult<Json, GetInfoFromUriError> {
     use http::header::HeaderValue;
     use mm2_net::transport::slurp_req_body;
 
@@ -234,7 +254,7 @@ async fn send_moralis_request(uri: &str) -> MmResult<Json, GetNftInfoError> {
 
     let (status, _header, body) = slurp_req_body(request).await?;
     if !status.is_success() {
-        return Err(MmError::new(GetNftInfoError::Transport(format!(
+        return Err(MmError::new(GetInfoFromUriError::Transport(format!(
             "Response !200 from {}: {}, {}",
             uri, status, body
         ))));
@@ -243,14 +263,14 @@ async fn send_moralis_request(uri: &str) -> MmResult<Json, GetNftInfoError> {
 }
 
 #[cfg(target_arch = "wasm32")]
-async fn send_moralis_request(uri: &str) -> MmResult<Json, GetNftInfoError> {
+async fn send_request_to_uri(uri: &str) -> MmResult<Json, GetInfoFromUriError> {
     use mm2_net::wasm_http::FetchRequest;
 
     macro_rules! try_or {
         ($exp:expr, $errtype:ident) => {
             match $exp {
                 Ok(x) => x,
-                Err(e) => return Err(MmError::new(GetNftInfoError::$errtype(ERRL!("{:?}", e)))),
+                Err(e) => return Err(MmError::new(GetInfoFromUriError::$errtype(ERRL!("{:?}", e)))),
             }
         };
     }
@@ -261,7 +281,7 @@ async fn send_moralis_request(uri: &str) -> MmResult<Json, GetNftInfoError> {
         .await;
     let (status_code, response_str) = try_or!(result, Transport);
     if !status_code.is_success() {
-        return Err(MmError::new(GetNftInfoError::Transport(ERRL!(
+        return Err(MmError::new(GetInfoFromUriError::Transport(ERRL!(
             "!200: {}, {}",
             status_code,
             response_str
@@ -288,4 +308,27 @@ pub(crate) async fn find_wallet_amount(
             token_id: token_id_req.to_string(),
         })?;
     Ok(nft.amount)
+}
+
+async fn try_get_uri_meta(token_uri: &Option<String>) -> MmResult<UriMeta, GetNftInfoError> {
+    match token_uri {
+        Some(token_uri) => {
+            if let Ok(response_meta) = send_request_to_uri(token_uri).await {
+                let uri_meta_res: UriMeta = serde_json::from_str(&response_meta.to_string())?;
+                Ok(uri_meta_res)
+            } else {
+                Ok(UriMeta::default())
+            }
+        },
+        None => Ok(UriMeta::default()),
+    }
+}
+
+fn get_tx_status(my_wallet: &str, to_address: &str) -> TransferStatus {
+    // if my_wallet == from_address && my_wallet == to_address it is incoming tx, so we can check just to_address.
+    if my_wallet.to_lowercase() == to_address.to_lowercase() {
+        TransferStatus::Receive
+    } else {
+        TransferStatus::Send
+    }
 }